IQC Fix group delete NPE

+++ src/main/java/com/stardust/easyassess/track/services/IQCPlanGroupServiceImpl.java
@@ -42,7 +42,9 @@ public class IQCPlanGroupServiceImpl extends AbstractEntityService<IQCPlanGroup>
         IQCPlanGroup group = this.get(id);
         if (group.getPlans() != null) {
             for (IQCPlan plan : group.getPlans()) {
-                iqcPlanRepository.delete(plan);
+                if (plan != null) {
+                    iqcPlanRepository.delete(plan);
+                }
             }
         }
